backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.1
apscheduler==3.10.4
cryptography==46.0.5
